Python/ML-TEST/ML_MK8-AI/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.0.9
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==41.0.3
cycler==0.11.0
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.5.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.12.1
idna==3.7
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.4.1
MarkupSafe==3.0.2
matplotlib==3.10.0
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
multidict==6.1.0
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.0
ply==3.11
propcache==0.3.1
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.10.1
pyOpenSSL==23.2.0
pyparsing==3.2.0
PyQt5_sip==12.13.0
PyQt5==5.15.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
scipy==1.15.2
setuptools==75.8.0
sip==6.7.12
six==1.17.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.1.0
tomli==2.0.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
win-inet-pton==1.1.0
wrapt==1.17.0
yarl==1.18.0
